requirements.txt
flask
qrcode
pillow
python-docx
